refactor(e2e): resolve pnpm/yarn corepack versions by major, not pin
Each smoke variant now passes a major to corepack (pnpm@10, pnpm@11,
yarn@1, yarn@4) instead of a pinned full version. Corepack resolves
the registry's current latest for that line on activation, so the
smoke tests track new patch/minor releases automatically.

Also consolidate the two identical yarn/pnpm corepack helpers into a
single `activatePackageManagerViaCorepack` that takes the manager name,
major, and an optional env override map.
