+++ e2e/src/smoke-tests/pnpm-10.spec.ts
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { smokeTest } from './smoke-test';
-import { activatePnpmViaCorepack } from './pnpm-corepack';
+import { activatePackageManagerViaCorepack } from './corepack';
 
 smokeTest('pnpm', {
   variant: 'pnpm-10',
-  setup: () => activatePnpmViaCorepack('10.33.2'),
+  setup: () => activatePackageManagerViaCorepack('pnpm', 10),
 });
