+++ e2e/src/smoke-tests/pnpm-11.spec.ts
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { smokeTest } from './smoke-test';
-import { activatePnpmViaCorepack } from './pnpm-corepack';
+import { activatePackageManagerViaCorepack } from './corepack';
 
 smokeTest('pnpm', {
   variant: 'pnpm-11',
-  setup: () => activatePnpmViaCorepack('11.0.4'),
+  setup: () => activatePackageManagerViaCorepack('pnpm', 11),
 });
