+++ e2e/src/smoke-tests/yarn.spec.ts
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { smokeTest } from './smoke-test';
+import { activatePackageManagerViaCorepack } from './corepack';
 
-// Classic yarn (1.x) — whichever version is on PATH. Paired with `yarn-4.spec.ts`
-// to cover both classic and berry code paths in create-nx-workspace.
-smokeTest('yarn', { variant: 'yarn-classic' });
+smokeTest('yarn', {
+  variant: 'yarn-classic',
+  setup: () => activatePackageManagerViaCorepack('yarn', 1),
+});
